Move `is_metric_improving()` and `TrendMetricNotSupportedError` from `metrics/domain/charts/line_with_shaded_section/information.py` -> `metrics.domain.common.information.py`
